requirements.txt
adafruit-circuitpython-neopixel>=6.3.0
adafruit-blinka>=8.20.0